port to 64 bit windows

+++ libraries/ObjectIO/ObjectIO/OS Windows/Windows_C_12/cCrossCall_121.h
@@ -50,28 +50,28 @@ extern void WinInitOs (Bool*,OS*);
 extern Bool WinCloseOs (OS);
 extern OS WinStartOsThread (OS);
 extern OS WinKillOsThread (OS);
-extern void WinKickOsThread (int,int,int,int,int,int,int,OS,int*,int*,int*,int*,int*,int*,int*,OS*);
+extern void WinKickOsThread (int,size_t,size_t,size_t,size_t,size_t,size_t,OS,int*,size_t*,size_t*,size_t*,size_t*,size_t*,size_t*,OS*);
 extern void KickCleanThread( CrossCallInfo *pcci );
 
-extern void SendMessageToClean( int mess, int p1,int p2,int p3, int p4,int p5,int p6 );
+extern void SendMessageToClean( int mess, size_t p1,size_t p2,size_t p3, size_t p4,size_t p5,size_t p6 );
 
 //	Shorthands for SendMessageToClean:
 #define SendMessage0ToClean(mess)                    SendMessageToClean((mess), 0,0,0,0,0,0)
-#define SendMessage1ToClean(mess, p1)                SendMessageToClean((mess), (int)(p1),0,0,0,0,0)
-#define SendMessage2ToClean(mess, p1,p2)             SendMessageToClean((mess), (int)(p1),(int)(p2),0,0,0,0)
-#define SendMessage3ToClean(mess, p1,p2,p3)          SendMessageToClean((mess), (int)(p1),(int)(p2),(int)(p3),0,0,0)
-#define SendMessage4ToClean(mess, p1,p2,p3,p4)       SendMessageToClean((mess), (int)(p1),(int)(p2),(int)(p3),(int)(p4),0,0)
-#define SendMessage5ToClean(mess, p1,p2,p3,p4,p5)    SendMessageToClean((mess), (int)(p1),(int)(p2),(int)(p3),(int)(p4),(int)(p5),0)
-#define SendMessage6ToClean(mess, p1,p2,p3,p4,p5,p6) SendMessageToClean((mess), (int)(p1),(int)(p2),(int)(p3),(int)(p4),(int)(p5),(int)(p6))
+#define SendMessage1ToClean(mess, p1)                SendMessageToClean((mess), (size_t)(p1),0,0,0,0,0)
+#define SendMessage2ToClean(mess, p1,p2)             SendMessageToClean((mess), (size_t)(p1),(size_t)(p2),0,0,0,0)
+#define SendMessage3ToClean(mess, p1,p2,p3)          SendMessageToClean((mess), (size_t)(p1),(size_t)(p2),(size_t)(p3),0,0,0)
+#define SendMessage4ToClean(mess, p1,p2,p3,p4)       SendMessageToClean((mess), (size_t)(p1),(size_t)(p2),(size_t)(p3),(size_t)(p4),0,0)
+#define SendMessage5ToClean(mess, p1,p2,p3,p4,p5)    SendMessageToClean((mess), (size_t)(p1),(size_t)(p2),(size_t)(p3),(size_t)(p4),(size_t)(p5),0)
+#define SendMessage6ToClean(mess, p1,p2,p3,p4,p5,p6) SendMessageToClean((mess), (size_t)(p1),(size_t)(p2),(size_t)(p3),(size_t)(p4),(size_t)(p5),(size_t)(p6))
 
 //	Prototypes of convenience functions that fill CrossCallInfo struct.
 extern CrossCallInfo *MakeReturn0Cci (CrossCallInfo * pcci);
-extern CrossCallInfo *MakeReturn1Cci (CrossCallInfo * pcci, int v);
-extern CrossCallInfo *MakeReturn2Cci (CrossCallInfo * pcci, int v1, int v2);
-extern CrossCallInfo *MakeReturn3Cci (CrossCallInfo * pcci, int v1, int v2, int v3);
-extern CrossCallInfo *MakeReturn4Cci (CrossCallInfo * pcci, int v1, int v2, int v3, int v4);
-extern CrossCallInfo *MakeReturn5Cci (CrossCallInfo * pcci, int v1, int v2, int v3, int v4, int v5);
-extern CrossCallInfo *MakeReturn6Cci (CrossCallInfo * pcci, int v1, int v2, int v3, int v4, int v5, int v6);
+extern CrossCallInfo *MakeReturn1Cci (CrossCallInfo * pcci, size_t v);
+extern CrossCallInfo *MakeReturn2Cci (CrossCallInfo * pcci, size_t v1, size_t v2);
+extern CrossCallInfo *MakeReturn3Cci (CrossCallInfo * pcci, size_t v1, size_t v2, size_t v3);
+extern CrossCallInfo *MakeReturn4Cci (CrossCallInfo * pcci, size_t v1, size_t v2, size_t v3, size_t v4);
+extern CrossCallInfo *MakeReturn5Cci (CrossCallInfo * pcci, size_t v1, size_t v2, size_t v3, size_t v4, size_t v5);
+extern CrossCallInfo *MakeReturn6Cci (CrossCallInfo * pcci, size_t v1, size_t v2, size_t v3, size_t v4, size_t v5, size_t v6);
 
 extern BOOL IsReturnCci( CrossCallInfo *pcci );
